refactor : Navigator => bottomTab 이름변경 및 features하위로 이동
