--- src/components/base/index.ts
@@ -1,1 +1,0 @@
-export {default as Navigator} from './items/Navigator';
