+++ src/components/features/bottomTab/index.ts
@@ -1,0 +1,1 @@
+export {default as BottomTab} from './BottomTab'
