+++ src/components/features/index.ts
@@ -1,2 +1,3 @@
 export {CustomCalendar} from './calendar'
 export {DayEventList} from './dayEventList'
+export {BottomTab} from './bottomTab'
